requirements.txt
streamlit
opencv-python-headless
numpy
requests
python-dotenv
Pillow
graphviz
Flask
